Move types declarion to types.py
