+++ mtpylon/__init__.py
@@ -1,9 +1,8 @@
 from .schema import Schema
-from .utils import long, int128, int256, double
+from .types import long, int128, int256, double
 
 __all__ = [
     'Schema',
-    'add',
     'long',
     'int128',
     'int256',
