+++ mtpylon/service_schema/constructors/client_dh_inner_data.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 from dataclasses import dataclass
 
-from mtpylon.utils import int128, long
+from mtpylon import long, int128
 
 
 @dataclass
