+++ mtpylon/service_schema/constructors/message.py
@@ -2,7 +2,7 @@
 from dataclasses import dataclass
 from typing import Any
 
-from mtpylon.utils import long
+from mtpylon import long
 
 
 @dataclass
