+++ mtpylon/service_schema/constructors/msgs_ack.py
@@ -2,7 +2,7 @@
 from dataclasses import dataclass
 from typing import List
 
-from mtpylon.utils import long
+from mtpylon import long
 
 
 @dataclass
