+++ mtpylon/service_schema/constructors/rpc_result.py
@@ -2,7 +2,7 @@
 from dataclasses import dataclass
 from typing import Any
 
-from mtpylon.utils import long
+from mtpylon import long
 
 
 @dataclass
