+++ mtpylon/service_schema/functions/set_client_DH_params_func.py
@@ -5,8 +5,8 @@ from contextvars import ContextVar
 
 from tgcrypto import ige256_decrypt  # type: ignore
 
-from mtpylon import Schema, long
-from mtpylon.utils import int128, int256, dump_integer_big_endian
+from mtpylon import Schema, long, int128, int256
+from mtpylon.utils import dump_integer_big_endian
 from mtpylon.crypto import AuthKey, KeyIvPair
 from mtpylon.contextvars import (
     new_nonce_var,
